エラーをなくす処理

+++ Game/PlayScene/Shadow/ShadowInformation.cpp
@@ -14,8 +14,8 @@ ShadowInformation::~ShadowInformation()
 void ShadowInformation::CreateDepthStencil()
 {
 	//		サイズ
-	RECT rect = { 0,0,LibrarySingleton::GetInstance()->GetScreenSize().x,
-				  LibrarySingleton::GetInstance()->GetScreenSize().y };
+	RECT rect = { 0,0,static_cast<int>(LibrarySingleton::GetInstance()->GetScreenSize().x),
+				  static_cast<int>(LibrarySingleton::GetInstance()->GetScreenSize().y) };
 
 	//		深度ステンシルの生成
 	m_depthStencil = std::make_unique<DepthStencil>(DXGI_FORMAT_D32_FLOAT);
@@ -29,8 +29,8 @@ void ShadowInformation::CreateDepthStencil()
 void ShadowInformation::CreateRenderTexture()
 {
 	//		サイズ
-	RECT rect = { 0,0,LibrarySingleton::GetInstance()->GetScreenSize().x,
-				  LibrarySingleton::GetInstance()->GetScreenSize().y };
+	RECT rect = { 0,0,static_cast<int>(LibrarySingleton::GetInstance()->GetScreenSize().x),
+				  static_cast<int>(LibrarySingleton::GetInstance()->GetScreenSize().y) };
 
 	//		深度ステンシルの生成
 	m_renderTexture = std::make_unique<DX::RenderTexture>(DXGI_FORMAT_R32_FLOAT);
